requirements.txt
tensorflow
streamlit
pandas
numpy
opencv-python
scikit-learn
matplotlib
Pillow
kaggle